new package structure
Maybe I'll like this organization more and stop tweaking it
